Abstract scheduling base
